backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.2
python-dotenv==1.1.0
bcrypt==4.2.1
PyJWT==2.10.1
email-validator==2.2.0
